Cleanup & minor improvements

+++ kos/src/kernel/core/dev/mktty.c
@@ -538,13 +538,11 @@ DEFINE_SYSCALL4(fd_t, mktty,
 		 *   - Write from to the display
 		 */
 		if unlikely(!IO_CANREAD(hkeyboard.h_mode))
-			THROW(E_INVALID_HANDLE_OPERATION,
-			      (unsigned int)keyboard,
+			THROW(E_INVALID_HANDLE_OPERATION, keyboard,
 			      E_INVALID_HANDLE_OPERATION_READ,
 			      hkeyboard.h_mode);
 		if unlikely(!IO_CANWRITE(hdisplay.h_mode))
-			THROW(E_INVALID_HANDLE_OPERATION,
-			      (unsigned int)display,
+			THROW(E_INVALID_HANDLE_OPERATION, display,
 			      E_INVALID_HANDLE_OPERATION_WRITE,
 			      hdisplay.h_mode);
